+++ lib/uuid.ex
@@ -1,6 +1,8 @@
 defmodule Moar.UUID do
   # @related [test](test/uuid_test.exs)
 
+  @moduledoc "UUID-related functions"
+
   @uuid_regex ~r/[0-9a-f]{2}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}/i
 
   @doc "Returns a regex that will match a UUID"